Split tools/mod.rs into cohesive submodules
tools/mod.rs was a 2400-line god-module. Break it into focused files,
mirroring the oauth/ layout, each well under 1k lines:

- params.rs       — tool parameter structs/enums + their schema tests
- handlers.rs     — the #[tool_router] impl (the 21 tools)
- gate.rs         — ToolGate (from the earlier refactor)
- values.rs       — Value extraction/sort/filter/strip helpers + tests
- validate.rs     — info-hash/label validation + source detection + tests
- errors.rs       — LLM-facing error enrichment
- plugin_watcher.rs — event→URI mapping + apply_label_state fan-out
- registry.rs     — tool registry (from the earlier refactor)

mod.rs now holds only the DelugeServer struct, constructor, plugin-watcher
task, the three stateful self-methods, and the ServerHandler impl (607
lines). Stateless helpers stay `pub(super)` methods on DelugeServer so call
sites are unchanged. The #[tool_router] macro's generated tool_router() is
private to handlers.rs, so a pub(super) build_tool_router() wrapper exposes
it to the constructor. Behaviour unchanged; all 40 tests pass.
